14 - TKInter/AanDeSlag/requirements.txt
requests==2.18.4
xmltodict==0.11.0
python-dateutil==2.7.3
